Bump react-native-macos version (#5089)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

This PR bumps `react-native-macos` version from 0.71 to 0.72. It also
features some minor changes that unifies `MacOSExample` and `Example`
apps.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

Tested with MacOSExample app.