Add boolean cast into `reducedMotion` (#5149)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

While working on refactor of `animationsManager.ts` file typescript
threw error that I cannot assign `{}` to `boolean`. Even though
refreshing window helped, we decided to keep this cast.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->
Checked that `reducedMotion` hook works.

---------

Co-authored-by: Tomasz Żelawski <40713406+tjzel@users.noreply.github.com>