Unify `Extrapolation` and `Extrapolate`. (#5141)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->
We have 2 similar objects that may be quite confusing:

<table>
<tr ><td>Extrapolate</td><td>Extrapolation</td></tr>

<tr><td>
Was valid in reanimated v1 


[LINK](https://docs.swmansion.com/react-native-reanimated/docs/1.x/nodes/interpolate/#interpolate)
</td>
<td>
Valid option according to our docs, 


[LINK](https://docs.swmansion.com/react-native-reanimated/docs/utilities/interpolate/#extrapolation-type-object--string)
</td></tr>

<tr><td>

```javascript
export const Extrapolate = {
  EXTEND: 'extend',
  CLAMP: 'clamp',
  IDENTITY: 'identity',
}
```
</td><td>

```javascript
export enum Extrapolation {
  IDENTITY = 'identity',
  CLAMP = 'clamp',
  EXTEND = 'extend',
}
```
</td></tr>
</table>

As you can see `Extrapolation` is now the "valid" name, and
"Extrapolate" is a legacy one. Therefore I mark it as `@deprecated`.
## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Aleksandra Cynk <aleksandra.cynk@swmansion.com>
Co-authored-by: Tomasz Żelawski <40713406+tjzel@users.noreply.github.com>