Fix article progress example on android (#5189)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

An error is thrown when you measure en element that is not mounted yet.
So I create a boolean flag which I set when component is ready to avoid
this error.

Probably we also should improve error log.

Current error log:
> `[Reanimated] The view with tag ${viewTag} returned an invalid
measurement response.`

New error log:
> `[Reanimated] The view with tag ${viewTag} returned an invalid
measurement response. Please make sure the view is currently rendered.`

Both logs are not perfect, since in our example our measured element is
a Text component, not a View one, however adding component type to log
is not easy.

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Aleksandra Cynk <aleksandra.cynk@swmansion.com>
Co-authored-by: Krzysztof Piaskowy <krzysztof.piaskowy@swmansion.com>