Improve spring config verification (#5188)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Each time a spring animation is triggered we check if provided config is
valid. When it isn't we skip animation and change the animated view to
its final position after waiting the time passed as `duration` (which
may be zero)

This change fixes the following bug:
* Warning when spring used with zero duration (which should be a valid
usage)
* No warning for negative mass

Additional benefits are:
* Invalid config warning informs what exactly is wrong
* Types are both more accurate and less verobse

Behaviour when duration is zero:


https://github.com/software-mansion/react-native-reanimated/assets/56199675/c16caa2a-9df0-4ba0-9890-4fe81b5b090d


<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Aleksandra Cynk <aleksandra.cynk@swmansion.com>