Refactor `animationsManager` on web (#5147)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

`animationsManager.ts` file grew pretty big during development of layout
animations. This PR aims to split it into multiple files to separate the
logic of DOM manipulation from the animations logic.
## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

Check that WebExample app works as it used to.