Update browserslist (#5258)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR updates browserslist to eliminate the following warning when
installing dependencies or running tests:

![Zrzut ekranu 2023-10-17 o 19 15
15](https://github.com/software-mansion/react-native-reanimated/assets/20516055/c734195e-49b4-4b4e-aee8-c6ef3c72f774)

I've run command `npx update-browserslist-db@latest`.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->