Remove plugin.js.map (#5257)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR removes `plugin.js.map` from the repository since ackchyually it
is not required when using Reanimated installed from branch (I hope so)
and also causes conflicts when there are plugin changes in two or more
PRs.

## Test plan

Install Reanimated from branch and check if it works