Fix ArticleProgressExample again (#5273)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR fixes ArticleProgressExample again. In current code,
`textRef.current` is passed as a copy so it never changes and thus the
example doesn't work. I've also decided to simplify the code and name
intermediate results of mathematical formulas used to calculate the
width of the progress bar.

cc @jfedak whom I've complained to that the example doesn't work on
macOS

cc @tjzel who never believed in `clamp`

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->