Fix stack trace on worklet factory crash (#5197)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Fixes #5165 

Since a worklet factory doesn't have it's counterpart in source code,
when it crashes we can get a useless stack trace. It can crash when a
variable that doesn't exists is referenced - since it's pulled into the
worklet's factory closure. This PR fixes the stack trace on such
crashes.

## Test plan

<details>
<summary>
Snippet that forces the factory crash
</summary>

```TSX
//
//
//
//
//
//

export default function EmptyExample() {
  function fooOOO() {
    'worklet';
    stackOrderAscendinggg;
  }

  return null;
}
```
</details>

| Before | After |
|--------|--------|
|![Simulator Screenshot - iPhone 14 Pro - 2023-10-06 at 19 23
48](https://github.com/software-mansion/react-native-reanimated/assets/40713406/2ebba9f7-874f-42bb-8bb5-27684acbb7ff)|![Simulator
Screenshot - iPhone 14 Pro - 2023-10-06 at 19 23
16](https://github.com/software-mansion/react-native-reanimated/assets/40713406/9aeddeb4-db67-4500-b3f8-1748c8d4f9d4)|