Create `commonTypes.ts` in `createAnimatedComponent` (#5282)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This pull requests moves types from `createAnimatedComponent` and
`utils.ts` to new `commonTypes.ts`. The reason for that is
`createAnimatedComponent`, `PropsFilter`, `JSUpdater` and
`InlinePropManager` depend on internals of one another, so circular
dependencies have to be avoided - also stuffing everything into `utils`
feels wrong.

This PR is required by #5278.

## Test plan

Standard TS test suite.