Fix lint staged (#5353)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
BEFORE:
* Lint-staged lints all files and prints all warnings and errors

NOW:
* Lint only staged files and shows both errors and warnings
* If the previous tasks success all files get lined, but only errors are
printed
<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->


Drawbacks:
Lint runs twice. It should not cause burdensome delays since the first
lint runs only for the staged files, so if you stage <10 files it runs
very fast.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Aleksandra Cynk <aleksandracynk@Aleksandras-MacBook-Pro-3.local>
Co-authored-by: Krzysztof Piaskowy <krzysztof.piaskowy@swmansion.com>